プレイ画面の背景を変更した

+++ avoiding/GameSource/Layer/PlayGameLayer.cpp
@@ -469,7 +469,7 @@ Vector3 PlayGameLayer::InitialPlacement(int mapSize_x, int mapSize_z, float offs
 // @ return : Vector3...座標                                                                       //
 // @ note	:                                                                                      //
 // ----------------------------------------------------------------------------------------------- // 
-Vector3 PlayGameLayer::CalculateElementNumber(Vector3 pos, int offset_x, int offset_z)
+Vector3 PlayGameLayer::CalculateElementNumber(Vector3 pos, float offset_x, float offset_z)
 {
 	// 戻り値の変数を定義
 	Vector3 result = Vector3(0, 0, 0);
